Reroute source of docs for dataframe methods
