+++ dask/dataframe/api.py
@@ -1,0 +1,4 @@
+from __future__ import annotations
+
+from dask.dataframe.dask_expr._groupby import GroupBy, SeriesGroupBy  # noqa: F401
+from dask.dataframe.dask_expr._rolling import Rolling  # noqa: F401
